refactor(RowMapper): QueryCallback의 이름 변경 및 인터페이스 구현체 생성
