+++ src/main/java/core/jdbc/RowMapper.java
@@ -3,6 +3,6 @@ package core.jdbc;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-public interface QueryCallback<T> {
+public interface RowMapper<T> {
     T mapping(ResultSet rs) throws SQLException;
 }
